second set of extra data for all loads
