+++ src/cards/compact/essentialLoad.ts
@@ -50,7 +50,7 @@ export class EssentialLoad {
 				data.essentialLoadCol1DynamicColour,
 				EssentialLoad.loadNameCol1(config),
 				data.essentialLoadCol1State,
-				data.essentialLoadCol1ExtraState,
+				LoadUtils.extraMode2(config) ? data.essentialLoadCol1Extra2State : data.essentialLoadCol1ExtraState,
 				data.essentialLoadCol1ToggleState,
 				config.load.auto_scale, data.decimalPlaces,
 				Load.column1, Load.row2,
@@ -116,7 +116,7 @@ export class EssentialLoad {
 				data.essentialLoadCol1DynamicColour,
 				EssentialLoad.loadNameCol1(config),
 				data.essentialLoadCol1State,
-				data.essentialLoadCol1ExtraState,
+				LoadUtils.extraMode2(config) ? data.essentialLoadCol1Extra2State : data.essentialLoadCol1ExtraState,
 				data.essentialLoadCol1ToggleState,
 				config.load.auto_scale, data.decimalPlaces,
 				Load.column2, Load.row2,
@@ -177,7 +177,7 @@ export class EssentialLoad {
 				data.essentialLoadCol1DynamicColour,
 				EssentialLoad.loadNameCol1(config),
 				data.essentialLoadCol1State,
-				data.essentialLoadCol1ExtraState,
+				LoadUtils.extraMode2(config) ? data.essentialLoadCol1ExtraState : data.essentialLoadCol1ExtraState,
 				data.essentialLoadCol1ToggleState,
 				config.load.auto_scale, data.decimalPlaces,
 				Load.column1, Load.row4,
@@ -224,7 +224,7 @@ export class EssentialLoad {
 				data.essentialLoadCol1DynamicColour,
 				EssentialLoad.loadNameCol1(config),
 				data.essentialLoadCol1State,
-				data.essentialLoadCol1ExtraState,
+				LoadUtils.extraMode2(config) ? data.essentialLoadCol1Extra2State : data.essentialLoadCol1ExtraState,
 				data.essentialLoadCol1ToggleState,
 				config.load.auto_scale, data.decimalPlaces,
 				Load.column2, Load.row4,
@@ -339,7 +339,7 @@ export class EssentialLoad {
 			data.essentialLoadCol1DynamicColour,
 			EssentialLoad.loadNameCol1(config),
 			data.essentialLoadCol1State,
-			data.essentialLoadCol1ExtraState,
+			LoadUtils.extraMode2(config) ? data.essentialLoadCol1Extra2State : data.essentialLoadCol1ExtraState,
 			data.essentialLoadCol1ToggleState,
 			config.load.auto_scale, data.decimalPlaces,
 			Load.column1, EssentialLoad.row(id),
@@ -355,7 +355,7 @@ export class EssentialLoad {
 			data.essentialLoadCol2DynamicColour,
 			EssentialLoad.loadNameCol2(config),
 			data.essentialLoadCol2State,
-			data.essentialLoadCol2ExtraState,
+			LoadUtils.extraMode2(config) ? data.essentialLoadCol2Extra2State : data.essentialLoadCol2ExtraState,
 			data.essentialLoadCol2ToggleState,
 			config.load.auto_scale, data.decimalPlaces,
 			Load.column2, EssentialLoad.row(id),
@@ -371,7 +371,7 @@ export class EssentialLoad {
 			data.essentialLoadCol3DynamicColour,
 			EssentialLoad.loadNameCol3(config),
 			data.essentialLoadCol3State,
-			data.essentialLoadCol3ExtraState,
+			LoadUtils.extraMode2(config) ? data.essentialLoadCol3Extra2State : data.essentialLoadCol3ExtraState,
 			data.essentialLoadCol3ToggleState,
 			config.load.auto_scale, data.decimalPlaces,
 			Load.column3, EssentialLoad.row(id),
@@ -387,7 +387,7 @@ export class EssentialLoad {
 			data.essentialLoadCol4DynamicColour,
 			EssentialLoad.loadNameCol4(config),
 			data.essentialLoadCol4State,
-			data.essentialLoadCol4ExtraState,
+			LoadUtils.extraMode2(config) ? data.essentialLoadCol4Extra2State : data.essentialLoadCol4ExtraState,
 			data.essentialLoadCol4ToggleState,
 			config.load.auto_scale, data.decimalPlaces,
 			Load.column4, EssentialLoad.row(id),
@@ -403,7 +403,7 @@ export class EssentialLoad {
 			data.essentialLoadCol5DynamicColour,
 			EssentialLoad.loadNameCol5(config),
 			data.essentialLoadCol5State,
-			data.essentialLoadCol5ExtraState,
+			LoadUtils.extraMode2(config) ? data.essentialLoadCol5Extra2State : data.essentialLoadCol5ExtraState,
 			data.essentialLoadCol5ToggleState,
 			config.load.auto_scale, data.decimalPlaces,
 			Load.column5, EssentialLoad.row(id),
@@ -419,7 +419,7 @@ export class EssentialLoad {
 			data.essentialLoadCol6DynamicColour,
 			EssentialLoad.loadNameCol6(config),
 			data.essentialLoadCol6State,
-			data.essentialLoadCol6ExtraState,
+			LoadUtils.extraMode2(config) ? data.essentialLoadCol6Extra2State : data.essentialLoadCol6ExtraState,
 			data.essentialLoadCol6ToggleState,
 			config.load.auto_scale, data.decimalPlaces,
 			Load.column6, EssentialLoad.row(id),
