+++ src/localize/languages/en.json
@@ -104,12 +104,15 @@
     "card_width": "Card Width (px or %)",
     "decimal_places": "Decimal Places",
     "decimal_places_energy": "Decimal Places (Daily Energy)",
+    "loads_extra_2_enabled": "Enable Loads Extra Mode Switch",
+    "loads_extra_2": "Extra 2 as Default",
+    "loads_extra_2_circle": "Show Extra Switch Circle",
     "wide_view_mode": "Wide View",
-    "align_grid":  "Align grid to left",
-    "align_load":  "Align load to right",
-    "viewbox_min_x":  "min-x",
+    "align_grid": "Align grid to left",
+    "align_load": "Align load to right",
+    "viewbox_min_x": "min-x",
     "viewbox_min_y": "min-y",
-    "viewbox_width":  "width",
+    "viewbox_width": "width",
     "viewbox_height": "height",
     "model": "Model",
     "autarky": "Autarky",
@@ -221,12 +224,10 @@
     "export_icon": "Export Icon",
     "disconnect_icon": "Disconnect Icon",
     "prepaid_unit_name": "Pre-paid Unit Name",
-
     "nonessential_name": "Name",
     "nonessential_icon": "Icon",
     "nonessential_daily_name": "Daily Name",
     "show_nonessential_daily": "Show Daily",
-
     "energy_cost_decimals": "Energy Cost Decimals",
     "show_remaining_energy": "Show Remaining Energy",
     "remaining_energy_to_shutdown": "Show Remaining Energy to Shutdown",
@@ -252,13 +253,11 @@
       "general": "General Options",
       "adv_viewbox": "Adv. ViewBox Options",
       "low_resources": "Low Resources Options",
-
+      "adv_options": "Advanced Options",
       "optional_ent": "Optional Entities",
-
       "inverter": "Inverter Options",
       "inv": "Inverter Entities",
       "inv_prog": "Programs Entities",
-
       "solar": "Solar Options",
       "solar_header": "Header Names",
       "solar_header_ent": "Header Entities",
@@ -269,7 +268,6 @@
       "pv_3_ent": "PV3 Entities",
       "pv_4_ent": "PV4 Entities",
       "pv_5_ent": "PV5 Entities",
-
       "battery": "Battery Options",
       "battery_ent": "Battery Entities",
       "battery_bank": "Battery Banks Options",
@@ -279,7 +277,6 @@
       "battery_bank_4_ent": "Battery Bank 4 Entities",
       "battery_bank_5_ent": "Battery Bank 5 Entities",
       "battery_bank_6_ent": "Battery Bank 6 Entities",
-
       "load": "Load Options",
       "load_ent": "Load Entities",
       "additional_loads": "Additional Loads Options",
@@ -297,13 +294,11 @@
       "ess_ld_4": "Column 4 Entities",
       "ess_ld_5": "Column 5 Entities",
       "ess_ld_6": "Column 6 Entities",
-      "additional_loads_view_mode" : "Additional Loads View Mode",
-
+      "additional_loads_view_mode": "Additional Loads View Mode",
       "aux_load": "Aux Load Options",
       "aux_load_ent": "Main Entities",
       "aux_load_row_1": "Item Options",
-      "aux_load_row_1_ent" : "Items Entities",
-
+      "aux_load_row_1_ent": "Items Entities",
       "grid": "Grid Options",
       "gri_ent": "Grid Entities",
       "grid_load": "Grid Load Options",
