+++ src/types.ts
@@ -86,6 +86,11 @@ export interface PowerFlowCardConfig extends LovelaceCardConfig {
 		refresh_interval: number,
 		animations: boolean,
 	},
+	adv_options: {
+		loads_extra_2_enabled: boolean
+		loads_extra_2: boolean
+		loads_extra_2_circle: boolean
+	}
 	inverter: {
 		modern: boolean;
 		colour: string;
@@ -751,34 +756,40 @@ export interface DataDto {
 	//new essential loads
 	additionalLoads: number,
 	essentialLoadCol1Icon: string[],
-	essentialLoadCol1State: CustomEntity[],
-	essentialLoadCol1ExtraState: CustomEntity[],
-	essentialLoadCol1ToggleState: CustomEntity[],
-	essentialLoadCol1DynamicColour: string[],
 	essentialLoadCol2Icon: string[],
-	essentialLoadCol2State: CustomEntity[],
-	essentialLoadCol2ExtraState: CustomEntity[],
-	essentialLoadCol2ToggleState: CustomEntity[],
-	essentialLoadCol2DynamicColour: string[],
 	essentialLoadCol3Icon: string[],
-	essentialLoadCol3State: CustomEntity[],
-	essentialLoadCol3ExtraState: CustomEntity[],
-	essentialLoadCol3ToggleState: CustomEntity[],
-	essentialLoadCol3DynamicColour: string[],
 	essentialLoadCol4Icon: string[],
-	essentialLoadCol4State: CustomEntity[],
-	essentialLoadCol4ExtraState: CustomEntity[],
-	essentialLoadCol4ToggleState: CustomEntity[],
-	essentialLoadCol4DynamicColour: string[],
 	essentialLoadCol5Icon: string[],
-	essentialLoadCol5State: CustomEntity[],
-	essentialLoadCol5ExtraState: CustomEntity[],
-	essentialLoadCol5ToggleState: CustomEntity[],
-	essentialLoadCol5DynamicColour: string[],
 	essentialLoadCol6Icon: string[],
+	essentialLoadCol1State: CustomEntity[],
+	essentialLoadCol2State: CustomEntity[],
+	essentialLoadCol3State: CustomEntity[],
+	essentialLoadCol4State: CustomEntity[],
+	essentialLoadCol5State: CustomEntity[],
 	essentialLoadCol6State: CustomEntity[],
+	essentialLoadCol1ExtraState: CustomEntity[],
+	essentialLoadCol2ExtraState: CustomEntity[],
+	essentialLoadCol3ExtraState: CustomEntity[],
+	essentialLoadCol4ExtraState: CustomEntity[],
+	essentialLoadCol5ExtraState: CustomEntity[],
 	essentialLoadCol6ExtraState: CustomEntity[],
+	essentialLoadCol1Extra2State: CustomEntity[],
+	essentialLoadCol2Extra2State: CustomEntity[],
+	essentialLoadCol3Extra2State: CustomEntity[],
+	essentialLoadCol4Extra2State: CustomEntity[],
+	essentialLoadCol5Extra2State: CustomEntity[],
+	essentialLoadCol6Extra2State: CustomEntity[],
+	essentialLoadCol1ToggleState: CustomEntity[],
+	essentialLoadCol2ToggleState: CustomEntity[],
+	essentialLoadCol3ToggleState: CustomEntity[],
+	essentialLoadCol4ToggleState: CustomEntity[],
+	essentialLoadCol5ToggleState: CustomEntity[],
 	essentialLoadCol6ToggleState: CustomEntity[],
+	essentialLoadCol1DynamicColour: string[],
+	essentialLoadCol2DynamicColour: string[],
+	essentialLoadCol3DynamicColour: string[],
+	essentialLoadCol4DynamicColour: string[],
+	essentialLoadCol5DynamicColour: string[],
 	essentialLoadCol6DynamicColour: string[],
 
 	enableTimer,
@@ -798,6 +809,7 @@ export interface DataDto {
 	nonessentialLoadIcon: string[],
 	nonessentialLoadState: CustomEntity[],
 	nonEssentialLoadExtraState: CustomEntity[],
+	nonEssentialLoadExtra2State: CustomEntity[],
 	nonEssentialLoadToggleState: CustomEntity[],
 	nonEssentialLoadMainDynamicColour: string,
 	nonEssentialLoadDynamicColour: string[],
@@ -854,6 +866,7 @@ export interface DataDto {
 	auxLoadDynamicColour: string[],
 	auxLoadState: CustomEntity[],
 	auxLoadExtraState: CustomEntity[],
+	auxLoadExtra2State: CustomEntity[],
 	auxLoadToggleState: CustomEntity[],
 
 	flowColour: string,
